Updated testing to be a bit friendlier.
The "test case" is the executeable that will load an SDL and SDL-TTF
environment, load a font, and render it to the screen until you tell
it to quit.

There is an awful lot that could go wrong outside the TTF system but
it's the simplest method of testing that the lib works. At least from
a truly basic functional point of view.

More advanced tests should probably be implemented for things like
checking the font and (text)ure attributes, styles, et al.